Refactor RelayMetadata into Relay model
