+++ Nos/Models/Nos.xcdatamodeld/Nos 10.xcdatamodel/.xccurrentversion
@@ -3,6 +3,6 @@
 <plist version="1.0">
 <dict>
 	<key>_XCCurrentVersionName</key>
-	<string>Nos 10.xcdatamodel</string>
+	<string>Nos.xcdatamodel</string>
 </dict>
 </plist>
