+++ Nos/Models/Relay+CoreDataProperties.swift
@@ -22,7 +22,16 @@ extension Relay {
     @NSManaged public var events: Set<Event>
     @NSManaged public var publishedEvents: Set<Event>
     @NSManaged public var shouldBePublishedEvents: Set<Event>
-    @NSManaged public var metadata: RelayMetadata?
+
+    // Metadata
+    @NSManaged public var name: String?
+    @NSManaged public var relayDescription: String?
+    @NSManaged public var supportedNIPs: [Int]?
+    @NSManaged public var pubkey: String?
+    @NSManaged public var contact: String?
+    @NSManaged public var software: String?
+    @NSManaged public var version: String?
+    @NSManaged public var metadataFetchedAt: Date?
 }
 
 // MARK: Generated accessors for authors
